Change `LOCATION` from arg to flag
According to CLI guidelines, "it's a bit more typing, but it makes it much clearer what is going on. It also makes it easier to make changes to how you accept input in the future. Sometimes when using args, it's impossible to add new input without breaking existing behavior or creating ambiguity".